Update config to support environment-specific variable setting (finally)
With the old config, running the test bot (which doesn't have access to the Infinite Nomic server) would cause it to throw errors trying to edit a channel it doesn't have access to. I would have to replace the channel ID values every time if I wanted those errors to not happen. Now I can have both the "debug" config values and the "production" config values in the config at the same time.