Implement saving current locale to settings file for persistence
I initially had the `initialize()` logic in the constructor, but it was leading to it being needlessly called multiple times for every single class.
Now it gets called once in the constructor (albeit in a kinda funky way with that `Locale(None)` bit, but I wanted to ensure that forgetting to include a base route was a runtime error)

The use of `db` in this commit is meant to be temporary. In the future, I would like these config values to be a part of a more variable name-driven config or settings class. I just still need to figure out exactly how I want to do that.